# public/swagger.yaml
...
openapi: 3.0.3
info:
  title: Task Titans API
  description: >
    Complete API documentation for Task Titans application with authentication, user management, and comprehensive
    workflows.


    ## Authentication Flows


    ### User Registration Flow

    **Step 1:** User submits registration data via `POST /user/`

    **Step 2:** System creates account and sends OTP verification email

    **Step 3:** User receives OTP code in email

    **Step 4:** User verifies email via `POST /auth/verify-email` with OTP

    **Step 5:** User account is verified and can now login


    ### Login Flow

    **Step 1:** User submits credentials via `POST /auth/login`

    **Step 2:** System validates credentials and returns JWT access token

    **Step 3:** User includes Bearer token in Authorization header for protected routes

    **Step 4:** Access protected endpoints like profile management


    ### Password Reset Flow

    **Step 1:** User requests password reset via `POST /auth/forget-password`

    **Step 2:** System sends OTP to user's email

    **Step 3:** User verifies OTP via `POST /auth/verify-email`

    **Step 4:** System returns reset token for password change

    **Step 5:** User submits new password via `POST /auth/reset-password` with token

    **Step 6:** Password is updated successfully


    ### Password Change Flow (Authenticated Users)

    **Step 1:** Authenticated user accesses `POST /auth/change-password`

...

...


    ### Rating Flow

...

...

...

...

...

...
  version: 1.0.0
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests

  # ---------------------------------------------------
  # 📦 SCHEMAS: Data models and definitions
  # ---------------------------------------------------
  schemas:
...
    CreateUserRequest:
      type: object
      required:
        - name
        - contact
        - email
        - password
        - location
      properties:
        name:
          type: string
          example: John Doe
          description: Full name of the user
        contact:
          type: string
          example: '+1234567890'
          description: Contact number
        email:
          type: string
          format: email
          example: john@example.com
          description: Email address (must be unique)
        password:
          type: string
          format: password
          minLength: 8
          example: StrongPassword123
          description: Password (minimum 8 characters)
        location:
          type: string
          example: New York, USA
          description: User's location
        profile:
          type: string
          description: Optional profile image URL

    # ✏️ Schema for UpdateProfileRequest
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          example: John Doe Updated
        contact:
          type: string
          example: '+1234567891'
        email:
          type: string
          format: email
          example: john.updated@example.com
        location:
          type: string
          example: Los Angeles, USA
        image:
          type: string
          format: binary
          description: Profile image file

    # ✏️ Schema for UserResponse
    UserResponse:
      type: object
      properties:
        _id:
          type: string
          example: 64f8a1b2c3d4e5f6a7b8c9d0
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: john@example.com
        contact:
          type: string
          example: '+1234567890'
        location:
          type: string
          example: New York, USA
        role:
          type: string
          enum:
            - SUPER_ADMIN
            - ADMIN
            - USER
          example: USER
        image:
          type: string
          example: https://i.ibb.co/z5YHLV9/profile.png
        status:
          type: string
          enum:
            - active
            - delete
          example: active
        verified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ✏️ Schema for LoginRequest
...
    RatingResponse:
      type: object
      properties:
        _id:
          type: string
          example: 64f8a1b2c3d4e5f6a7b8c9d2
        taskId:
          type: object
          properties:
            _id:
              type: string
              example: 64f8a1b2c3d4e5f6a7b8c9d0
            title:
              type: string
              example: Fix the car engine
            description:
              type: string
              example: Engine requires a full service
        raterId:
          type: object
          properties:
            _id:
              type: string
              example: 64f8a1b2c3d4e5f6a7b8c9d3
            name:
              type: string
              example: John Doe
            email:
              type: string
              example: john@example.com
            image:
              type: string
              example: https://i.ibb.co/z5YHLV9/profile.png
        ratedUserId:
          type: object
          properties:
            _id:
              type: string
              example: 64f8a1b2c3d4e5f6a7b8c9d1
            name:
              type: string
              example: Jane Smith
            email:
              type: string
              example: jane@example.com
            image:
              type: string
              example: https://i.ibb.co/z5YHLV9/profile.png
        rating:
          type: number
          example: 5
        comment:
          type: string
          example: Excellent work and great communication!
        ratingType:
          type: string
          enum:
            - task_completion
            - communication
            - quality
            - timeliness
          example: task_completion
        status:
          type: string
          enum:
            - active
            - deleted
          example: active
        createdAt:
          type: string
          format: date-time
          example: '2025-08-23T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2025-08-23T12:00:00Z'

    # ✏️ Schema for RatingStatsResponse
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        statusCode:
          type: number
          example: 200
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
          description: Response data (varies by endpoint)

    # ✏️ Schema for ErrorResponse
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        statusCode:
          type: number
          example: 400
        message:
          type: string
          example: Error message
        errorMessages:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string

    # ✏️ Schema for Bid
...
paths:
...
  /ratings/user/{userId}:
    get:
      tags:
        - Rating System
      summary: Get ratings for a specific user
      description: |-
        Retrieve all ratings received by a specific user

        ⚡ This GET endpoint is cacheable. Consider caching responses.
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: User ID
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: ratingType
          in: query
          schema:
            type: string
            enum:
              - task_completion
              - communication
              - quality
              - timeliness
      responses:
        '200':
          description: User ratings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User ratings retrieved successfully
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RatingResponse'
                  meta:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      total:
                        type: integer
                        example: 15
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Internal server error

  # ---------------------------------------------------
  # 🟢 GET /ratings/user/{userId}/stats — Get user rating statistics
  # ---------------------------------------------------
...
  /user/:
    post:
      tags:
        - User Management
      summary: Create a new user account
      description: |-
        Register a new user account. After successful registration:
        - User receives an OTP verification email
        - Account is created but not verified
        - User must verify email before login

        ⏳ This method modifies data. Apply appropriate rate limiting to protect the server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      message:
                        example: User created successfully
                      data:
                        $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/profile:
    # ---------------------------------------------------
    # 🟢 GET /user/profile — Get current user profile
    # ---------------------------------------------------
    get:
      tags:
        - User Management
      summary: Get current user profile
      description: |-
        Retrieve the authenticated user's profile information

        ⚡ This GET endpoint is cacheable. Consider caching responses.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      message:
                        example: Profile data retrieved successfully
                      data:
                        $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    # ---------------------------------------------------
    # 🟡 PATCH /user/profile — Update user profile
    # ---------------------------------------------------
    patch:
      tags:
        - User Management
      summary: Update user profile
      description: |-
        Update the authenticated user's profile information.
        Supports file upload for profile image.

        ⏳ This method modifies data. Apply appropriate rate limiting to protect the server.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                data:
                  type: string
                  description: JSON string containing profile update data
                  example: '{"name":"Updated Name","contact":"+1234567891"}'
                image:
                  type: string
                  format: binary
                  description: Profile image file
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      message:
                        example: Profile updated successfully
                      data:
                        $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ===================================================
  # BID MODULE — Bid Management related endpoints
  # ===================================================

  # ---------------------------------------------------
  # 🔵 POST /bids — Create a new bid
  # ---------------------------------------------------
...
security:
  - BearerAuth: []