# src/app/modules/user/user.route.ts
import express, { NextFunction, Request, Response } from 'express';
import { USER_ROLES } from '../../../enums/user';
import auth from '../../middlewares/auth';
import fileUploadHandler from '../../middlewares/fileUploadHandler';
import validateRequest from '../../middlewares/validateRequest';
import { UserController } from './user.controller';
import { UserValidation } from './user.validation';

const router = express.Router();

router
  .route('/profile')
  // GET /profile → Get current user's profile
  .get(auth(USER_ROLES.ADMIN, USER_ROLES.USER), UserController.getUserProfile)

  // PATCH /profile → Update user profile
  .patch(
    auth(USER_ROLES.SUPER_ADMIN, USER_ROLES.ADMIN, USER_ROLES.USER),
    fileUploadHandler(), // Handle profile picture upload
    (req: Request, res: Response, next: NextFunction) => {
      if (req.body.data) {
        // Parse and validate JSON data from form-data
        req.body = UserValidation.updateUserZodSchema.parse(
          JSON.parse(req.body.data)
        );
      }
      return UserController.updateProfile(req, res, next);
    }
  );

router
  .route('/')
  // POST / → Create a new user
  .post(
    validateRequest(UserValidation.createUserZodSchema),
    UserController.createUser
  );

export const UserRoutes = router;